DeployApiTest/django_app_docker/ApiTest/requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
colorlog==4.0.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
inflection==0.3.1
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
packaging==19.2
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
